ai_model.py
# ai_model.py

import torch
from torch import nn
from config import AI_MODEL_CONFIG

class SocialAuthenticityModel(nn.Module):
    def __init__(self):
        super(SocialAuthenticityModel, self).__init__()
        # Define your model architecture here
        # This is just a placeholder
        self.layer1 = nn.Linear(100, 50)
        self.layer2 = nn.Linear(50, 10)
        self.layer3 = nn.Linear(10, 1)

    def forward(self, x):
        x = torch.relu(self.layer1(x))
        x = torch.relu(self.layer2(x))
        x = torch.sigmoid(self.layer3(x))
        return x

class AIModel:
    def __init__(self):
        self.model = SocialAuthenticityModel()
        self.model_path = AI_MODEL_CONFIG['MODEL_PATH']
        self.load_model()

    def load_model(self):
        try:
            self.model.load_state_dict(torch.load(self.model_path))
            self.model.eval()
        except Exception as e:
            print(f"Error loading the model: {e}")

    def predict(self, user_data):
        # Preprocess user_data and convert it to tensor
        # This is just a placeholder
        user_data_tensor = torch.tensor(user_data)
        with torch.no_grad():
            prediction = self.model(user_data_tensor)
        return prediction.item()

ai_model_instance = AIModel()
